move EventSourcing.Persistence.SqlServer DI code into main EventSourcing.Persistence.SqlServer package
