+++ EventSourcing.Persistence/SqlServer/Persistence.SqlServer.IntegrationTests/Persistence.SqlServer.IntegrationTests.csproj
@@ -27,7 +27,6 @@
       <ProjectReference Include="..\..\..\EventSourcing.Serialization\Extensions.DependencyInjection.Serialization.NewtonsoftJson\Extensions.DependencyInjection.Serialization.NewtonsoftJson.csproj" />
       <ProjectReference Include="..\..\Persistence.IntegrationTests.Base\Persistence.IntegrationTests.Base.csproj" />
       <ProjectReference Include="..\Extensions.DatabaseMigrations.Persistence.SqlServer.DbUp\Extensions.DatabaseMigrations.Persistence.SqlServer.DbUp.csproj" />
-      <ProjectReference Include="..\Extensions.DependencyInjection.Persistence.SqlServer\Extensions.DependencyInjection.Persistence.SqlServer.csproj" />
       <ProjectReference Include="..\Persistence.SqlServer\Persistence.SqlServer.csproj" />
     </ItemGroup>
